Fixed text colors
All text colors now print when asked to and look like the color they are
supposed to be. Not in multishell mode, though. It's staying in black
and white until I get the balls to touch the code. Not that it's
unmaintainable, it's well written and I don't want to break it.

Oh yeah, and it's no longer 5588:00 PM on day 1.