Improved Rednet support
This seems complete, but sometimes when I try to transfer files manually insted of using qft, it fails
inexplicably due to a serialization error. Because of that, I don't consider it complete, but qft works
fine and the failures are rare. In case there is a failure, check rednet.log, which will be generated in
/CraftOS. Careful, the log only logs the latest transmission to conserve disk space.