Fixed some fs bugs
I specifically fixed bugs pertaining to the deletion of files in rom and the deletion of mount points.
It should be slightly more stable now.

NEW BUGS:
Shell doesn't print the welcome message in colour.
Shell doesn't print the text prompt in colour.
If the screen scrolls all the way to the bottom, colours will stop working until the screen is cleared.