Updated rom files to version 1.80pr1
This one was tricky. I had to properely implement the colours api due to
the new palette thing, and I've probably programmed it wrong due to
there being little to no documentation on this beta release. Even now,
it isn't complete, as the window API seems to have gotten all the
colours flip turned upside down. And I'd like to take a minute, just sit
right there....

I need a nap and a coffee break.

Also, I have no idea how to find a system's timezone, so only UTC mode
is implemented.
Also note that I left out the command and turtle stuff again because I
won't be using it.